// pronti-frontend-main/front/src/app/modulos/cuenta-contable/servicios/dto/cuenta-contable.response-dto.ts
import {AbstractResponseDto} from "../../../../abstract/dto/abstract-response.dto";
import {TipoCuentaEnum} from '../../../../enums/tipo-cuenta.enum';
import {TipoMovimientoEnum} from '../../../../enums/tipo-movimiento.enum';
import {PeriodoContableResponseDto} from '../../../periodo-contable/servicios/dto/periodo-contable.response-dto';
import {AgenciaResponseDto} from "../../../agencia/servicios/dto/agencia.response-dto";
import {RubrosRolResponseDto} from "../../../rubros-rol/servicios/dto/rubros-rol.response-dto";

export interface CuentaContableResponseDto extends AbstractResponseDto {
  // nombre?: string;
  idPeriodoContable?: PeriodoContableResponseDto;
  idAgencia?: AgenciaResponseDto;
  idRubrosRol?: RubrosRolResponseDto;
  nombre?: string;
  identificador?: number;
  nivel?: number;
  idNivel1?: number;
  idNivel2?: number;
  idNivel3?: number;
  idNivel4?: number;
  idNivel5?: number;
  tipoCuenta?: TipoCuentaEnum;
  movimiento?: TipoMovimientoEnum;
  anteriorDebito?: number;
  anteriorCredito?: number;
  anteriorSaldo?: number;
  eneroDebito?: number;
  eneroCredito?: number;
  eneroSaldo?: number;
  febreroDebito?: number;
  febreroCredito?: number;
  febreroSaldo?: number;
  marzoDebito?: number;
  marzoCredito?: number;
  marzoSaldo?: number;
  abrilDebito?: number;
  abrilCredito?: number;
  abrilSaldo?: number;
  mayoDebito?: number;
  mayoCredito?: number;
  mayoSaldo?: number;
  junioDebito?: number;
  junioCredito?: number;
  junioSaldo?: number;
  julioDebito?: number;
  julioCredito?: number;
  julioSaldo?: number;
  agostoDebito?: number;
  agostoCredito?: number;
  agostoSaldo?: number;
  septiembreDebito?: number;
  septiembreCredito?: number;
  septiembreSaldo?: number;
  octubreDebito?: number;
  octubreCredito?: number;
  octubreSaldo?: number;
  noviembreDebito?: number;
  noviembreCredito?: number;
  noviembreSaldo?: number;
  diciembreDebito?: number;
  diciembreCredito?: number;
  diciembreSaldo?: number;
  actualDebito?: number;
  actualCredito?: number;
  actualSaldo?: number;
}